Use depth8 to detect free TT entries
Check depth8 instead of key16 to determine if a TTEntry is free.
To make sure that depth8 is never zero in used entries, MAX_PLY
and the constant added to depth on saving are adjusted by 1.

Using key16 here is not perfect because there are keys whose highest
16 bits are all zero. They cannot be retrieved back and may hide
other entries.

Bench: 3447564